- Refactored code:   - Moved FieldMatch annotation to validation package.   - Changed HTTP status for RegistrationException to CONFLICT.   - Removed unnecessary @Mapping in UserMapper.   - Used @Getter and @Setter instead of @Data for User entity.   - Improved exception message in UserServiceImpl.
