+++ src/main/java/mate/academy/repository/UserRepository.java
@@ -1,11 +1,10 @@
 package mate.academy.repository;
 
-import java.util.Optional;
 import mate.academy.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
+    boolean existsByEmail(String email);
 }
