+++ src/main/java/mate/academy/service/impl/UserServiceImpl.java
@@ -19,12 +19,12 @@ public class UserServiceImpl implements UserService {
     @Override
     public UserResponseDto register(UserRegistrationRequestDto requestDto)
             throws RegistrationException {
-        if (userRepository.findByEmail(requestDto.getEmail()).isPresent()) {
-            throw new RegistrationException("Email already in use");
+        if (userRepository.existsByEmail(requestDto.getEmail())) {
+            throw new RegistrationException("Email already in use: " + requestDto.getEmail());
         }
 
         User user = userMapper.toModel(requestDto);
-        User savedUser = userRepository.save(user);
-        return userMapper.toDto(savedUser);
+        userRepository.save(user);
+        return userMapper.toDto(user);
     }
 }
